+++ src/convertCsvToXlsx.ts
@@ -14,8 +14,10 @@ import type {APIParameters} from './convertCsvToXlsx.types';
 export function convertCsvToXlsx(
   source: string,
   destination: string,
-  {sheetName = '', overwrite = false}: APIParameters = {},
+  apiParameters: APIParameters = {},
 ) {
+  const {sheetName = '', overwrite = false} = apiParameters;
+
   // sanity checks
   if (typeof source !== 'string' || typeof destination !== 'string') {
     throw new Error(
@@ -49,7 +51,7 @@ export function convertCsvToXlsx(
 
   // csv parser options
   const csvOptions: Options = {
-    columns: true,
+    columns: false,
     delimiter: ',',
     ltrim: true,
     rtrim: true,
@@ -62,7 +64,7 @@ export function convertCsvToXlsx(
   const wb = xlsx.utils.book_new();
 
   // insert the records as a sheet
-  const ws = xlsx.utils.json_to_sheet(records);
+  const ws = xlsx.utils.json_to_sheet(records, {skipHeader: true});
   xlsx.utils.book_append_sheet(wb, ws, sheetName);
 
   // write the xlsx workbook to destination